v2.7

+++ visidata/loaders/sqlite.py
@@ -21,7 +21,9 @@ class SqliteSheet(Sheet):
 
     def conn(self):
         import sqlite3
-        return sqlite3.connect(str(self.resolve()))
+        con = sqlite3.connect(str(self.resolve()))
+        con.text_factory = lambda s, enc=self.options.encoding, encerrs=self.options.encoding_errors: s.decode(enc, encerrs)
+        return con
 
     def execute(self, conn, sql, parms=None):
         parms = parms or []
@@ -42,11 +44,11 @@ class SqliteSheet(Sheet):
                 return int if f == '0' else float
             return anytype
 
+        self.rowidColumn = None
         with self.conn() as conn:
             tblname = self.tableName
             if not isinstance(self, SqliteIndexSheet):
                 self.columns = []
-                self.addColumn(ColumnItem('rowid', 0, type=int, width=0))
                 for r in self.execute(conn, 'PRAGMA TABLE_XINFO("%s")' % tblname):
                     colnum, colname, coltype, nullable, defvalue, colkey, *_ = r
                     c = ColumnItem(colname, colnum+1, type=parse_sqlite_type(coltype))
@@ -55,10 +57,14 @@ class SqliteSheet(Sheet):
                     if colkey:
                         self.setKeys([c])
 
-            try:
+                if 'WITHOUT ROWID' not in self.row[5]: # SQL used to create table
+                    self.rowidColumn = ColumnItem('rowid', 0, type=int, width=0)
+                    self.addColumn(self.rowidColumn, index=0)
+
+            if self.rowidColumn:
                 r = self.execute(conn, 'SELECT rowid, * FROM "%s"' % tblname)
-            except sqlite3.OperationalError:
-                vd.error('tables WITHOUT ROWID not supported')
+            else:
+                r = self.execute(conn, 'SELECT NULL, * FROM "%s"' % tblname)
             yield from Progress(r, total=r.rowcount-1)
 
     @asyncthread
@@ -83,7 +89,6 @@ class SqliteSheet(Sheet):
             return vals
 
         with self.conn() as conn:
-            wherecols = [self.columns[0]] # self.column("rowid")
             for r in adds.values():
                 cols = self.visibleCols
                 sql = 'INSERT INTO "%s" ' % self.tableName
@@ -94,6 +99,10 @@ class SqliteSheet(Sheet):
                     vd.warning('not all rows inserted') # f'{res.rowcount}/{res.arraysize} rows inserted'
 
             for row, rowmods in mods.values():
+                if not self.rowidColumn:
+                    vd.warning('cannot modify rows in tables without rowid')
+                    break
+                wherecols = [self.rowidColumn]
                 sql = 'UPDATE "%s" SET ' % self.tableName
                 sql += ', '.join('%s=?' % c.name for c, _ in rowmods.items())
                 sql += ' WHERE %s' % ' AND '.join('"%s"=?' % c.name for c in wherecols)
@@ -105,6 +114,11 @@ class SqliteSheet(Sheet):
                     vd.warning('not all rows updated') # f'{res.rowcount}/{res.arraysize} rows updated'
 
             for row in dels.values():
+                if not self.rowidColumn:
+                    vd.warning('cannot delete rows in tables without rowid')
+                    break
+
+                wherecols = [self.rowidColumn]
                 sql = 'DELETE FROM "%s" ' % self.tableName
                 sql += ' WHERE %s' % ' AND '.join('"%s"=?' % c.name for c in wherecols)
                 wherevals=list(Column.calcValue(c, row) for c in wherecols)
@@ -124,6 +138,7 @@ class SqliteIndexSheet(SqliteSheet, IndexSheet):
     savesToSource = True
     defer = True
     def iterload(self):
+        self.addColumn(Column('sql', width=0, getter=lambda c,r:r.row[5]))
         for row in SqliteSheet.iterload(self):
             if row[1] != 'index':
                 tblname = row[2]
@@ -171,6 +186,7 @@ class SqliteQuerySheet(SqliteSheet):
 def save_sqlite(vd, p, *vsheets):
     import sqlite3
     conn = sqlite3.connect(str(p))
+    conn.text_factory = lambda s, enc=vsheets[0].options.encoding: s.decode(enc)
     conn.row_factory = sqlite3.Row
     c = conn.cursor()
 
@@ -204,7 +220,7 @@ def save_sqlite(vd, p, *vsheets):
                 elif not isinstance(v, (int, float, str)):
                     v = col.getDisplayValue(r)
                 sqlvals.append(v)
-            sql = 'INSERT INTO "%s" (%s) VALUES (%s)' % (tblname, ','.join(c.name for c in vs.visibleCols), ','.join('?' for v in sqlvals))
+            sql = 'INSERT INTO "%s" (%s) VALUES (%s)' % (tblname, ','.join(f'"{c.name}"' for c in vs.visibleCols), ','.join('?' for v in sqlvals))
             c.execute(sql, sqlvals)
 
     conn.commit()
